Add basic unstyled sturcture to the site
This change adds the basic components to the site. It attempts to try
to break out each part of the site into its own React Component. I am
still learning React, so this structure it likely to be clean up.

Nothing works or talks to real data. Also I have made efforts to try
not to hard code any of the text or component data. Right now all of
the rendering state is defined with plains JS Maps and Constants.

Just the index page works.