Add week of action events to the site
This change adds the current week of action to the static site. This
is managed by a JSON file and not action network. In order to have
something by the end of the week I have taken a short cut and just
hard coded the events to the site. This should make it more usable.