Handle date formating
This change adds nicer date handling logic to the website. Action
Network’s API returns these date objects as if they are in UTC,
however they are actually in local time. This change strips off the
timestamp of the date objects and then uses the local browser timezone
to correctly format the time.